feat(test): add test for v2
