+++ src/test/kotlin/it/pagopa/ecommerce/helpdesk/dataproviders/mongo/EcommerceForTransactionV2DataProviderTest.kt
@@ -1,16 +1,19 @@
 package it.pagopa.ecommerce.helpdesk.dataproviders.mongo
 
-import it.pagopa.ecommerce.commons.documents.v1.TransactionAuthorizationRequestData
-import it.pagopa.ecommerce.commons.documents.v1.TransactionClosureData
-import it.pagopa.ecommerce.commons.documents.v1.TransactionEvent
-import it.pagopa.ecommerce.commons.documents.v1.TransactionUserReceiptData
-import it.pagopa.ecommerce.commons.domain.v1.TransactionWithUserReceiptOk
-import it.pagopa.ecommerce.commons.domain.v1.pojos.*
-import it.pagopa.ecommerce.commons.generated.server.model.AuthorizationResultDto
+import it.pagopa.ecommerce.commons.documents.v2.TransactionAuthorizationRequestData as TransactionAuthorizationRequestDataV2
+import it.pagopa.ecommerce.commons.documents.v2.TransactionClosureData as TransactionClosureDataV2
+import it.pagopa.ecommerce.commons.documents.v2.TransactionEvent as TransactionEventV2
+import it.pagopa.ecommerce.commons.documents.v2.TransactionUserReceiptData as TransactionUserReceiptDataV2
+import it.pagopa.ecommerce.commons.domain.v2.pojos.BaseTransactionExpired as BaseTransactionExpiredV2
+import it.pagopa.ecommerce.commons.domain.v2.pojos.BaseTransactionWithClosureError as BaseTransactionWithClosureErrorV2
+import it.pagopa.ecommerce.commons.domain.v2.pojos.BaseTransactionWithCompletedAuthorization as BaseTransactionWithCompletedAuthorizationV2
+import it.pagopa.ecommerce.commons.domain.v2.pojos.BaseTransactionWithRefundRequested as BaseTransactionWithRefundRequestedV2
+import it.pagopa.ecommerce.commons.domain.v2.pojos.BaseTransactionWithRequestedUserReceipt as BaseTransactionWithRequestedUserReceiptV2
+import it.pagopa.ecommerce.commons.domain.v2.pojos.BaseTransactionWithUserReceipt as BaseTransactionWithUserReceiptV2
+import it.pagopa.ecommerce.commons.generated.npg.v1.dto.OperationResultDto
 import it.pagopa.ecommerce.commons.generated.server.model.TransactionStatusDto
-import it.pagopa.ecommerce.commons.v1.TransactionTestUtils
+import it.pagopa.ecommerce.commons.v2.TransactionTestUtils as TransactionTestUtilsV2
 import it.pagopa.ecommerce.helpdesk.HelpdeskTestUtils
-import it.pagopa.ecommerce.helpdesk.exceptions.InvalidSearchCriteriaException
 import it.pagopa.generated.ecommerce.helpdesk.model.*
 import java.time.ZonedDateTime
 import org.junit.jupiter.api.AfterAll
@@ -24,7 +27,7 @@ import reactor.core.publisher.Flux
 import reactor.core.publisher.Mono
 import reactor.test.StepVerifier
 
-class EcommerceTransactionDataProviderTest {
+class EcommerceForTransactionV2DataProviderTest {
 
     companion object {
         val testedStatuses: MutableSet<TransactionStatusDto> = HashSet()
@@ -57,849 +60,37 @@ class EcommerceTransactionDataProviderTest {
             transactionsViewRepository,
             transactionsEventStoreRepository
         )
-
     @Test
-    fun `should count total transactions by rptId successfully`() {
-        val searchCriteria = HelpdeskTestUtils.buildSearchRequestByRptId()
-        given(transactionsViewRepository.countTransactionsWithRptId(searchCriteria.rptId))
-            .willReturn(Mono.just(2))
-        StepVerifier.create(ecommerceTransactionDataProvider.totalRecordCount(searchCriteria))
-            .expectNext(2)
-            .verifyComplete()
-    }
-
-    @Test
-    fun `should count total transactions by paymentToken successfully`() {
-        val searchCriteria = HelpdeskTestUtils.buildSearchRequestByPaymentToken()
-        given(
-                transactionsViewRepository.countTransactionsWithPaymentToken(
-                    searchCriteria.paymentToken
-                )
-            )
-            .willReturn(Mono.just(2))
-        StepVerifier.create(ecommerceTransactionDataProvider.totalRecordCount(searchCriteria))
-            .expectNext(2)
-            .verifyComplete()
-    }
-
-    @Test
-    fun `should count total transactions by transactionId successfully`() {
-        val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
-        given(transactionsViewRepository.existsById(searchCriteria.transactionId))
-            .willReturn(Mono.just(true))
-        StepVerifier.create(ecommerceTransactionDataProvider.totalRecordCount(searchCriteria))
-            .expectNext(1)
-            .verifyComplete()
-    }
-
-    @Test
-    fun `should handle no transactions found by rptId successfully`() {
-        val searchCriteria = HelpdeskTestUtils.buildSearchRequestByRptId()
-        given(transactionsViewRepository.countTransactionsWithRptId(searchCriteria.rptId))
-            .willReturn(Mono.just(0))
-        StepVerifier.create(ecommerceTransactionDataProvider.totalRecordCount(searchCriteria))
-            .expectNext(0)
-            .verifyComplete()
-    }
-
-    @Test
-    fun `should handle no transaction found by paymentToken successfully`() {
-        val searchCriteria = HelpdeskTestUtils.buildSearchRequestByPaymentToken()
-        given(
-                transactionsViewRepository.countTransactionsWithPaymentToken(
-                    searchCriteria.paymentToken
-                )
-            )
-            .willReturn(Mono.just(0))
-        StepVerifier.create(ecommerceTransactionDataProvider.totalRecordCount(searchCriteria))
-            .expectNext(0)
-            .verifyComplete()
-    }
-
-    @Test
-    fun `should handle no transaction found by transactionId successfully`() {
-        val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
-        given(transactionsViewRepository.existsById(searchCriteria.transactionId))
-            .willReturn(Mono.just(false))
-        StepVerifier.create(ecommerceTransactionDataProvider.totalRecordCount(searchCriteria))
-            .expectNext(0)
-            .verifyComplete()
-    }
-
-    @Test
-    fun `should return error for search by email as invalid search criteria`() {
-        val searchCriteria = HelpdeskTestUtils.buildSearchRequestByUserMail("test@test.it")
-        StepVerifier.create(ecommerceTransactionDataProvider.totalRecordCount(searchCriteria))
-            .expectError(InvalidSearchCriteriaException::class.java)
-            .verify()
-    }
-
-    @Test
-    fun `should return error for search by user fiscal code as invalid search criteria`() {
-        val searchCriteria = HelpdeskTestUtils.buildSearchRequestByUserFiscalCode("fiscal code")
-        StepVerifier.create(ecommerceTransactionDataProvider.totalRecordCount(searchCriteria))
-            .expectError(InvalidSearchCriteriaException::class.java)
-            .verify()
-    }
-
-    @Test
-    fun `should return error for search by unknown search criteria`() {
-        val searchCriteria: HelpDeskSearchTransactionRequestDto = mock()
-        given(searchCriteria.type).willReturn("UNKNOWN")
-        StepVerifier.create(ecommerceTransactionDataProvider.totalRecordCount(searchCriteria))
-            .expectError(InvalidSearchCriteriaException::class.java)
-            .verify()
-    }
-
-    @Test
-    fun `should map successfully transaction data into response searching by rptId for NOTIFIED_OK transaction`() {
-        val searchCriteria = HelpdeskTestUtils.buildSearchRequestByRptId()
-        val pageSize = 100
-        val pageNumber = 0
-        val transactionView =
-            TransactionTestUtils.transactionDocument(
-                TransactionStatusDto.NOTIFIED_OK,
-                ZonedDateTime.now()
-            )
-        val transactionUserReceiptData =
-            TransactionTestUtils.transactionUserReceiptData(TransactionUserReceiptData.Outcome.OK)
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
-        val authorizationRequestedEvent =
-            TransactionTestUtils.transactionAuthorizationRequestedEvent()
-        val authorizedEvent = TransactionTestUtils.transactionAuthorizationCompletedEvent()
-        val closureSentEvent =
-            TransactionTestUtils.transactionClosedEvent(TransactionClosureData.Outcome.KO)
-        val addUserReceiptEvent =
-            TransactionTestUtils.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
-        val userReceiptAddErrorEvent =
-            TransactionTestUtils.transactionUserReceiptAddErrorEvent(addUserReceiptEvent.data)
-        val userReceiptAddedEvent =
-            TransactionTestUtils.transactionUserReceiptAddedEvent(userReceiptAddErrorEvent.data)
-        val events =
-            listOf(
-                transactionActivatedEvent,
-                authorizationRequestedEvent,
-                authorizedEvent,
-                closureSentEvent,
-                addUserReceiptEvent,
-                userReceiptAddErrorEvent,
-                userReceiptAddedEvent
-            )
-                as List<TransactionEvent<Any>>
-        val baseTransaction =
-            TransactionTestUtils.reduceEvents(*events.toTypedArray())
-                as TransactionWithUserReceiptOk
-        given(
-                transactionsViewRepository
-                    .findTransactionsWithRptIdPaginatedOrderByCreationDateDesc(
-                        rptId = searchCriteria.rptId,
-                        skip = pageSize * pageNumber,
-                        limit = pageSize
-                    )
-            )
-            .willReturn(Flux.just(transactionView))
-        given(
-                transactionsEventStoreRepository.findByTransactionIdOrderByCreationDateAsc(
-                    transactionView.transactionId
-                )
-            )
-            .willReturn(Flux.fromIterable(events))
-        val amount = baseTransaction.paymentNotices.sumOf { it.transactionAmount.value }
-        val fee = baseTransaction.transactionAuthorizationRequestData.fee
-        val totalAmount = amount.plus(fee)
-        val expected =
-            listOf(
-                TransactionResultDto()
-                    .userInfo(UserInfoDto().authenticationType("GUEST").notificationEmail(""))
-                    .transactionInfo(
-                        TransactionInfoDto()
-                            .creationDate(baseTransaction.creationDate.toOffsetDateTime())
-                            .status("Confermato")
-                            .statusDetails(null)
-                            .eventStatus(
-                                it.pagopa.generated.ecommerce.helpdesk.model.TransactionStatusDto
-                                    .valueOf(transactionView.status.toString())
-                            )
-                            .amount(amount)
-                            .fee(fee)
-                            .grandTotal(totalAmount)
-                            .rrn(baseTransaction.transactionAuthorizationCompletedData.rrn)
-                            .authorizationCode(
-                                baseTransaction.transactionAuthorizationCompletedData
-                                    .authorizationCode
-                            )
-                            .paymentMethodName(
-                                baseTransaction.transactionAuthorizationRequestData
-                                    .paymentMethodName
-                            )
-                            .brand(
-                                baseTransaction.transactionAuthorizationRequestData.brand!!
-                                    .toString()
-                            )
-                    )
-                    .paymentInfo(
-                        PaymentInfoDto()
-                            .origin(baseTransaction.clientId.toString())
-                            .details(
-                                baseTransaction.paymentNotices.map {
-                                    PaymentDetailInfoDto()
-                                        .subject(it.transactionDescription.value)
-                                        .rptId(it.rptId.value)
-                                        .idTransaction(baseTransaction.transactionId.value())
-                                        .paymentToken(it.paymentToken.value)
-                                        .creditorInstitution(
-                                            baseTransaction.transactionUserReceiptData
-                                                .receivingOfficeName
-                                        )
-                                        .paFiscalCode(it.transferList[0].paFiscalCode)
-                                }
-                            )
-                    )
-                    .pspInfo(
-                        PspInfoDto()
-                            .pspId(baseTransaction.transactionAuthorizationRequestData.pspId)
-                            .businessName(
-                                baseTransaction.transactionAuthorizationRequestData.pspBusinessName
-                            )
-                            .idChannel(
-                                baseTransaction.transactionAuthorizationRequestData.pspChannelCode
-                            )
-                    )
-                    .product(ProductDto.ECOMMERCE)
-            )
-        StepVerifier.create(
-                ecommerceTransactionDataProvider.findResult(
-                    searchParams = searchCriteria,
-                    skip = pageSize * pageNumber,
-                    limit = pageSize
-                )
-            )
-            .expectNext(expected)
-            .verifyComplete()
-    }
-
-    @Test
-    fun `should map successfully transaction data into response searching by paymentToken for NOTIFIED_OK transaction`() {
-        val searchCriteria = HelpdeskTestUtils.buildSearchRequestByPaymentToken()
-        val pageSize = 100
-        val pageNumber = 0
-        val transactionView =
-            TransactionTestUtils.transactionDocument(
-                TransactionStatusDto.NOTIFIED_OK,
-                ZonedDateTime.now()
-            )
-        val transactionUserReceiptData =
-            TransactionTestUtils.transactionUserReceiptData(TransactionUserReceiptData.Outcome.OK)
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
-        val authorizationRequestedEvent =
-            TransactionTestUtils.transactionAuthorizationRequestedEvent()
-        val authorizedEvent = TransactionTestUtils.transactionAuthorizationCompletedEvent()
-        val closureSentEvent =
-            TransactionTestUtils.transactionClosedEvent(TransactionClosureData.Outcome.KO)
-        val addUserReceiptEvent =
-            TransactionTestUtils.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
-        val userReceiptAddErrorEvent =
-            TransactionTestUtils.transactionUserReceiptAddErrorEvent(addUserReceiptEvent.data)
-        val userReceiptAddedEvent =
-            TransactionTestUtils.transactionUserReceiptAddedEvent(userReceiptAddErrorEvent.data)
-        val events =
-            listOf(
-                transactionActivatedEvent,
-                authorizationRequestedEvent,
-                authorizedEvent,
-                closureSentEvent,
-                addUserReceiptEvent,
-                userReceiptAddErrorEvent,
-                userReceiptAddedEvent
-            )
-                as List<TransactionEvent<Any>>
-        val baseTransaction =
-            TransactionTestUtils.reduceEvents(*events.toTypedArray())
-                as TransactionWithUserReceiptOk
-        given(
-                transactionsViewRepository
-                    .findTransactionsWithPaymentTokenPaginatedOrderByCreationDateDesc(
-                        paymentToken = searchCriteria.paymentToken,
-                        skip = pageSize * pageNumber,
-                        limit = pageSize
-                    )
-            )
-            .willReturn(Flux.just(transactionView))
-        given(
-                transactionsEventStoreRepository.findByTransactionIdOrderByCreationDateAsc(
-                    transactionView.transactionId
-                )
-            )
-            .willReturn(Flux.fromIterable(events))
-        val amount = baseTransaction.paymentNotices.sumOf { it.transactionAmount.value }
-        val fee = baseTransaction.transactionAuthorizationRequestData.fee
-        val totalAmount = amount.plus(fee)
-        val expected =
-            listOf(
-                TransactionResultDto()
-                    .userInfo(UserInfoDto().authenticationType("GUEST").notificationEmail(""))
-                    .transactionInfo(
-                        TransactionInfoDto()
-                            .creationDate(baseTransaction.creationDate.toOffsetDateTime())
-                            .status("Confermato")
-                            .statusDetails(null)
-                            .eventStatus(
-                                it.pagopa.generated.ecommerce.helpdesk.model.TransactionStatusDto
-                                    .valueOf(transactionView.status.toString())
-                            )
-                            .amount(amount)
-                            .fee(fee)
-                            .grandTotal(totalAmount)
-                            .rrn(baseTransaction.transactionAuthorizationCompletedData.rrn)
-                            .authorizationCode(
-                                baseTransaction.transactionAuthorizationCompletedData
-                                    .authorizationCode
-                            )
-                            .paymentMethodName(
-                                baseTransaction.transactionAuthorizationRequestData
-                                    .paymentMethodName
-                            )
-                            .brand(
-                                baseTransaction.transactionAuthorizationRequestData.brand!!
-                                    .toString()
-                            )
-                    )
-                    .paymentInfo(
-                        PaymentInfoDto()
-                            .origin(baseTransaction.clientId.toString())
-                            .details(
-                                baseTransaction.paymentNotices.map {
-                                    PaymentDetailInfoDto()
-                                        .subject(it.transactionDescription.value)
-                                        .rptId(it.rptId.value)
-                                        .idTransaction(baseTransaction.transactionId.value())
-                                        .paymentToken(it.paymentToken.value)
-                                        .creditorInstitution(
-                                            baseTransaction.transactionUserReceiptData
-                                                .receivingOfficeName
-                                        )
-                                        .paFiscalCode(it.transferList[0].paFiscalCode)
-                                }
-                            )
-                    )
-                    .pspInfo(
-                        PspInfoDto()
-                            .pspId(baseTransaction.transactionAuthorizationRequestData.pspId)
-                            .businessName(
-                                baseTransaction.transactionAuthorizationRequestData.pspBusinessName
-                            )
-                            .idChannel(
-                                baseTransaction.transactionAuthorizationRequestData.pspChannelCode
-                            )
-                    )
-                    .product(ProductDto.ECOMMERCE)
-            )
-        StepVerifier.create(
-                ecommerceTransactionDataProvider.findResult(
-                    searchParams = searchCriteria,
-                    skip = pageSize * pageNumber,
-                    limit = pageSize
-                )
-            )
-            .expectNext(expected)
-            .verifyComplete()
-    }
-
-    @Test
-    fun `should map successfully transaction data into response searching by transaction id for NOTIFIED_OK transaction`() {
+    fun `should map successfully transaction V2 data into response searching by transaction id for transaction in AUTHORIZATION_COMPLETED KO state`() {
         val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
         val pageSize = 100
         val pageNumber = 0
         val transactionView =
-            TransactionTestUtils.transactionDocument(
-                TransactionStatusDto.NOTIFIED_OK,
-                ZonedDateTime.now()
-            )
-        val transactionUserReceiptData =
-            TransactionTestUtils.transactionUserReceiptData(TransactionUserReceiptData.Outcome.OK)
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
-        val authorizationRequestedEvent =
-            TransactionTestUtils.transactionAuthorizationRequestedEvent()
-        val authorizedEvent = TransactionTestUtils.transactionAuthorizationCompletedEvent()
-        val closureSentEvent =
-            TransactionTestUtils.transactionClosedEvent(TransactionClosureData.Outcome.KO)
-        val addUserReceiptEvent =
-            TransactionTestUtils.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
-        val userReceiptAddErrorEvent =
-            TransactionTestUtils.transactionUserReceiptAddErrorEvent(addUserReceiptEvent.data)
-        val userReceiptAddedEvent =
-            TransactionTestUtils.transactionUserReceiptAddedEvent(userReceiptAddErrorEvent.data)
-        val events =
-            listOf(
-                transactionActivatedEvent,
-                authorizationRequestedEvent,
-                authorizedEvent,
-                closureSentEvent,
-                addUserReceiptEvent,
-                userReceiptAddErrorEvent,
-                userReceiptAddedEvent
-            )
-                as List<TransactionEvent<Any>>
-        val baseTransaction =
-            TransactionTestUtils.reduceEvents(*events.toTypedArray())
-                as TransactionWithUserReceiptOk
-        given(transactionsViewRepository.findById(searchCriteria.transactionId))
-            .willReturn(Mono.just(transactionView))
-        given(
-                transactionsEventStoreRepository.findByTransactionIdOrderByCreationDateAsc(
-                    transactionView.transactionId
-                )
-            )
-            .willReturn(Flux.fromIterable(events))
-        val amount = baseTransaction.paymentNotices.sumOf { it.transactionAmount.value }
-        val fee = baseTransaction.transactionAuthorizationRequestData.fee
-        val totalAmount = amount.plus(fee)
-        val expected =
-            listOf(
-                TransactionResultDto()
-                    .userInfo(UserInfoDto().authenticationType("GUEST").notificationEmail(""))
-                    .transactionInfo(
-                        TransactionInfoDto()
-                            .creationDate(baseTransaction.creationDate.toOffsetDateTime())
-                            .status("Confermato")
-                            .statusDetails(null)
-                            .eventStatus(
-                                it.pagopa.generated.ecommerce.helpdesk.model.TransactionStatusDto
-                                    .valueOf(transactionView.status.toString())
-                            )
-                            .amount(amount)
-                            .fee(fee)
-                            .grandTotal(totalAmount)
-                            .rrn(baseTransaction.transactionAuthorizationCompletedData.rrn)
-                            .authorizationCode(
-                                baseTransaction.transactionAuthorizationCompletedData
-                                    .authorizationCode
-                            )
-                            .paymentMethodName(
-                                baseTransaction.transactionAuthorizationRequestData
-                                    .paymentMethodName
-                            )
-                            .brand(
-                                baseTransaction.transactionAuthorizationRequestData.brand!!
-                                    .toString()
-                            )
-                    )
-                    .paymentInfo(
-                        PaymentInfoDto()
-                            .origin(baseTransaction.clientId.toString())
-                            .details(
-                                baseTransaction.paymentNotices.map {
-                                    PaymentDetailInfoDto()
-                                        .subject(it.transactionDescription.value)
-                                        .rptId(it.rptId.value)
-                                        .idTransaction(baseTransaction.transactionId.value())
-                                        .paymentToken(it.paymentToken.value)
-                                        .creditorInstitution(
-                                            baseTransaction.transactionUserReceiptData
-                                                .receivingOfficeName
-                                        )
-                                        .paFiscalCode(it.transferList[0].paFiscalCode)
-                                }
-                            )
-                    )
-                    .pspInfo(
-                        PspInfoDto()
-                            .pspId(baseTransaction.transactionAuthorizationRequestData.pspId)
-                            .businessName(
-                                baseTransaction.transactionAuthorizationRequestData.pspBusinessName
-                            )
-                            .idChannel(
-                                baseTransaction.transactionAuthorizationRequestData.pspChannelCode
-                            )
-                    )
-                    .product(ProductDto.ECOMMERCE)
-            )
-        StepVerifier.create(
-                ecommerceTransactionDataProvider.findResult(
-                    searchParams = searchCriteria,
-                    skip = pageSize,
-                    limit = pageNumber
-                )
-            )
-            .consumeNextWith {
-                assertEquals(expected, it)
-                testedStatuses.add(
-                    TransactionStatusDto.valueOf(it[0].transactionInfo.eventStatus.toString())
-                )
-            }
-            .verifyComplete()
-    }
-
-    @Test
-    fun `should return error for invalid search by user fiscal code`() {
-        val searchCriteria = HelpdeskTestUtils.buildSearchRequestByUserFiscalCode("fiscal code")
-        val pageSize = 100
-        val pageNumber = 0
-
-        StepVerifier.create(
-                ecommerceTransactionDataProvider.findResult(
-                    searchParams = searchCriteria,
-                    skip = pageSize,
-                    limit = pageNumber
-                )
-            )
-            .expectError(InvalidSearchCriteriaException::class.java)
-            .verify()
-    }
-
-    @Test
-    fun `should return error for invalid search by user email`() {
-        val searchCriteria = HelpdeskTestUtils.buildSearchRequestByUserMail("email@email.it")
-        val pageSize = 100
-        val pageNumber = 0
-
-        StepVerifier.create(
-                ecommerceTransactionDataProvider.findResult(
-                    searchParams = searchCriteria,
-                    skip = pageSize,
-                    limit = pageNumber
-                )
-            )
-            .expectError(InvalidSearchCriteriaException::class.java)
-            .verify()
-    }
-
-    @Test
-    fun `should return error for invalid search by unknown search criteria`() {
-        val searchCriteria: HelpDeskSearchTransactionRequestDto = mock()
-        val pageSize = 100
-        val pageNumber = 0
-        given(searchCriteria.type).willReturn("UNKNOWN")
-        StepVerifier.create(
-                ecommerceTransactionDataProvider.findResult(
-                    searchParams = searchCriteria,
-                    skip = pageSize,
-                    limit = pageNumber
-                )
-            )
-            .expectError(InvalidSearchCriteriaException::class.java)
-            .verify()
-    }
-
-    @Test
-    fun `should map successfully transaction data into response searching by transaction id for transaction in ACTIVATED state`() {
-        val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
-        val pageSize = 100
-        val pageNumber = 0
-        val transactionView =
-            TransactionTestUtils.transactionDocument(
-                TransactionStatusDto.ACTIVATED,
-                ZonedDateTime.now()
-            )
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
-        val events = listOf(transactionActivatedEvent) as List<TransactionEvent<Any>>
-        val baseTransaction = TransactionTestUtils.reduceEvents(*events.toTypedArray())
-        given(transactionsViewRepository.findById(searchCriteria.transactionId))
-            .willReturn(Mono.just(transactionView))
-        given(
-                transactionsEventStoreRepository.findByTransactionIdOrderByCreationDateAsc(
-                    transactionView.transactionId
-                )
-            )
-            .willReturn(Flux.fromIterable(events))
-        val amount = baseTransaction.paymentNotices.sumOf { it.transactionAmount.value }
-        val fee = 0
-        val totalAmount = amount.plus(fee)
-        val expected =
-            listOf(
-                TransactionResultDto()
-                    .userInfo(UserInfoDto().authenticationType("GUEST").notificationEmail(""))
-                    .transactionInfo(
-                        TransactionInfoDto()
-                            .creationDate(baseTransaction.creationDate.toOffsetDateTime())
-                            .status("Cancellato")
-                            .statusDetails(null)
-                            .eventStatus(
-                                it.pagopa.generated.ecommerce.helpdesk.model.TransactionStatusDto
-                                    .valueOf(transactionView.status.toString())
-                            )
-                            .amount(amount)
-                            .fee(fee)
-                            .grandTotal(totalAmount)
-                            .rrn(null)
-                            .authorizationCode(null)
-                            .paymentMethodName(null)
-                            .brand(null)
-                    )
-                    .paymentInfo(
-                        PaymentInfoDto()
-                            .origin(baseTransaction.clientId.toString())
-                            .details(
-                                baseTransaction.paymentNotices.map {
-                                    PaymentDetailInfoDto()
-                                        .subject(it.transactionDescription.value)
-                                        .rptId(it.rptId.value)
-                                        .idTransaction(baseTransaction.transactionId.value())
-                                        .paymentToken(it.paymentToken.value)
-                                        .creditorInstitution(null)
-                                        .paFiscalCode(it.transferList[0].paFiscalCode)
-                                }
-                            )
-                    )
-                    .pspInfo(PspInfoDto().pspId(null).businessName(null).idChannel(null))
-                    .product(ProductDto.ECOMMERCE)
-            )
-        StepVerifier.create(
-                ecommerceTransactionDataProvider.findResult(
-                    searchParams = searchCriteria,
-                    skip = pageSize,
-                    limit = pageNumber
-                )
-            )
-            .consumeNextWith {
-                assertEquals(expected, it)
-                testedStatuses.add(
-                    TransactionStatusDto.valueOf(it[0].transactionInfo.eventStatus.toString())
-                )
-            }
-            .verifyComplete()
-    }
-
-    @Test
-    fun `should map successfully transaction data into response searching by transaction id for transaction in AUTHORIZATION_REQUESTED state`() {
-        val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
-        val pageSize = 100
-        val pageNumber = 0
-        val transactionView =
-            TransactionTestUtils.transactionDocument(
-                TransactionStatusDto.AUTHORIZATION_REQUESTED,
-                ZonedDateTime.now()
-            )
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
-        val transactionAuthorizationRequestedEvent =
-            TransactionTestUtils.transactionAuthorizationRequestedEvent(
-                TransactionAuthorizationRequestData.PaymentGateway.XPAY
-            )
-        val events =
-            listOf(transactionActivatedEvent, transactionAuthorizationRequestedEvent)
-                as List<TransactionEvent<Any>>
-        val baseTransaction =
-            TransactionTestUtils.reduceEvents(*events.toTypedArray())
-                as BaseTransactionWithRequestedAuthorization
-        given(transactionsViewRepository.findById(searchCriteria.transactionId))
-            .willReturn(Mono.just(transactionView))
-        given(
-                transactionsEventStoreRepository.findByTransactionIdOrderByCreationDateAsc(
-                    transactionView.transactionId
-                )
-            )
-            .willReturn(Flux.fromIterable(events))
-        val amount = baseTransaction.paymentNotices.sumOf { it.transactionAmount.value }
-        val fee = baseTransaction.transactionAuthorizationRequestData.fee
-        val totalAmount = amount.plus(fee)
-        val expected =
-            listOf(
-                TransactionResultDto()
-                    .userInfo(UserInfoDto().authenticationType("GUEST").notificationEmail(""))
-                    .transactionInfo(
-                        TransactionInfoDto()
-                            .creationDate(baseTransaction.creationDate.toOffsetDateTime())
-                            .status("Cancellato")
-                            .statusDetails(null)
-                            .eventStatus(
-                                it.pagopa.generated.ecommerce.helpdesk.model.TransactionStatusDto
-                                    .valueOf(transactionView.status.toString())
-                            )
-                            .amount(amount)
-                            .fee(fee)
-                            .grandTotal(totalAmount)
-                            .rrn(null)
-                            .authorizationCode(null)
-                            .paymentMethodName(
-                                baseTransaction.transactionAuthorizationRequestData
-                                    .paymentMethodName
-                            )
-                            .brand(
-                                baseTransaction.transactionAuthorizationRequestData.brand!!
-                                    .toString()
-                            )
-                    )
-                    .paymentInfo(
-                        PaymentInfoDto()
-                            .origin(baseTransaction.clientId.toString())
-                            .details(
-                                baseTransaction.paymentNotices.map {
-                                    PaymentDetailInfoDto()
-                                        .subject(it.transactionDescription.value)
-                                        .rptId(it.rptId.value)
-                                        .idTransaction(baseTransaction.transactionId.value())
-                                        .paymentToken(it.paymentToken.value)
-                                        .creditorInstitution(null)
-                                        .paFiscalCode(it.transferList[0].paFiscalCode)
-                                }
-                            )
-                    )
-                    .pspInfo(
-                        PspInfoDto()
-                            .pspId(baseTransaction.transactionAuthorizationRequestData.pspId)
-                            .businessName(
-                                baseTransaction.transactionAuthorizationRequestData.pspBusinessName
-                            )
-                            .idChannel(
-                                baseTransaction.transactionAuthorizationRequestData.pspChannelCode
-                            )
-                    )
-                    .product(ProductDto.ECOMMERCE)
-            )
-        StepVerifier.create(
-                ecommerceTransactionDataProvider.findResult(
-                    searchParams = searchCriteria,
-                    skip = pageSize,
-                    limit = pageNumber
-                )
-            )
-            .consumeNextWith {
-                assertEquals(expected, it)
-                testedStatuses.add(
-                    TransactionStatusDto.valueOf(it[0].transactionInfo.eventStatus.toString())
-                )
-            }
-            .verifyComplete()
-    }
-
-    @Test
-    fun `should map successfully transaction data into response searching by transaction id for transaction in AUTHORIZATION_COMPLETED OK state`() {
-        val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
-        val pageSize = 100
-        val pageNumber = 0
-        val transactionView =
-            TransactionTestUtils.transactionDocument(
+            TransactionTestUtilsV2.transactionDocument(
                 TransactionStatusDto.AUTHORIZATION_COMPLETED,
                 ZonedDateTime.now()
             )
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
+        val transactionActivatedEvent = TransactionTestUtilsV2.transactionActivateEvent()
         val transactionAuthorizationRequestedEvent =
-            TransactionTestUtils.transactionAuthorizationRequestedEvent(
-                TransactionAuthorizationRequestData.PaymentGateway.XPAY
+            TransactionTestUtilsV2.transactionAuthorizationRequestedEvent(
+                TransactionAuthorizationRequestDataV2.PaymentGateway.XPAY
             )
         val transactionAuthorizationCompletedEvent =
-            TransactionTestUtils.transactionAuthorizationCompletedEvent(AuthorizationResultDto.OK)
+            TransactionTestUtilsV2.transactionAuthorizationCompletedEvent(
+                TransactionTestUtilsV2.npgTransactionGatewayAuthorizationData(
+                    OperationResultDto.FAILED
+                )
+            )
         val events =
             listOf(
                 transactionActivatedEvent,
                 transactionAuthorizationRequestedEvent,
                 transactionAuthorizationCompletedEvent
             )
-                as List<TransactionEvent<Any>>
+                as List<TransactionEventV2<Any>>
         val baseTransaction =
-            TransactionTestUtils.reduceEvents(*events.toTypedArray())
-                as BaseTransactionWithCompletedAuthorization
-        given(transactionsViewRepository.findById(searchCriteria.transactionId))
-            .willReturn(Mono.just(transactionView))
-        given(
-                transactionsEventStoreRepository.findByTransactionIdOrderByCreationDateAsc(
-                    transactionView.transactionId
-                )
-            )
-            .willReturn(Flux.fromIterable(events))
-        val amount = baseTransaction.paymentNotices.sumOf { it.transactionAmount.value }
-        val fee = baseTransaction.transactionAuthorizationRequestData.fee
-        val totalAmount = amount.plus(fee)
-        val expected =
-            listOf(
-                TransactionResultDto()
-                    .userInfo(UserInfoDto().authenticationType("GUEST").notificationEmail(""))
-                    .transactionInfo(
-                        TransactionInfoDto()
-                            .creationDate(baseTransaction.creationDate.toOffsetDateTime())
-                            .status("Confermato")
-                            .statusDetails(null)
-                            .eventStatus(
-                                it.pagopa.generated.ecommerce.helpdesk.model.TransactionStatusDto
-                                    .valueOf(transactionView.status.toString())
-                            )
-                            .amount(amount)
-                            .fee(fee)
-                            .grandTotal(totalAmount)
-                            .rrn(baseTransaction.transactionAuthorizationCompletedData.rrn)
-                            .authorizationCode(
-                                baseTransaction.transactionAuthorizationCompletedData
-                                    .authorizationCode
-                            )
-                            .paymentMethodName(
-                                baseTransaction.transactionAuthorizationRequestData
-                                    .paymentMethodName
-                            )
-                            .brand(
-                                baseTransaction.transactionAuthorizationRequestData.brand!!
-                                    .toString()
-                            )
-                    )
-                    .paymentInfo(
-                        PaymentInfoDto()
-                            .origin(baseTransaction.clientId.toString())
-                            .details(
-                                baseTransaction.paymentNotices.map {
-                                    PaymentDetailInfoDto()
-                                        .subject(it.transactionDescription.value)
-                                        .rptId(it.rptId.value)
-                                        .idTransaction(baseTransaction.transactionId.value())
-                                        .paymentToken(it.paymentToken.value)
-                                        .creditorInstitution(null)
-                                        .paFiscalCode(it.transferList[0].paFiscalCode)
-                                }
-                            )
-                    )
-                    .pspInfo(
-                        PspInfoDto()
-                            .pspId(baseTransaction.transactionAuthorizationRequestData.pspId)
-                            .businessName(
-                                baseTransaction.transactionAuthorizationRequestData.pspBusinessName
-                            )
-                            .idChannel(
-                                baseTransaction.transactionAuthorizationRequestData.pspChannelCode
-                            )
-                    )
-                    .product(ProductDto.ECOMMERCE)
-            )
-        StepVerifier.create(
-                ecommerceTransactionDataProvider.findResult(
-                    searchParams = searchCriteria,
-                    skip = pageSize,
-                    limit = pageNumber
-                )
-            )
-            .consumeNextWith {
-                assertEquals(expected, it)
-                testedStatuses.add(
-                    TransactionStatusDto.valueOf(it[0].transactionInfo.eventStatus.toString())
-                )
-            }
-            .verifyComplete()
-    }
-
-    @Test
-    fun `should map successfully transaction data into response searching by transaction id for transaction in AUTHORIZATION_COMPLETED KO state`() {
-        val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
-        val pageSize = 100
-        val pageNumber = 0
-        val transactionView =
-            TransactionTestUtils.transactionDocument(
-                TransactionStatusDto.AUTHORIZATION_COMPLETED,
-                ZonedDateTime.now()
-            )
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
-        val transactionAuthorizationRequestedEvent =
-            TransactionTestUtils.transactionAuthorizationRequestedEvent(
-                TransactionAuthorizationRequestData.PaymentGateway.XPAY
-            )
-        val transactionAuthorizationCompletedEvent =
-            TransactionTestUtils.transactionAuthorizationCompletedEvent(AuthorizationResultDto.KO)
-        val events =
-            listOf(
-                transactionActivatedEvent,
-                transactionAuthorizationRequestedEvent,
-                transactionAuthorizationCompletedEvent
-            )
-                as List<TransactionEvent<Any>>
-        val baseTransaction =
-            TransactionTestUtils.reduceEvents(*events.toTypedArray())
-                as BaseTransactionWithCompletedAuthorization
+            TransactionTestUtilsV2.reduceEvents(*events.toTypedArray())
+                as BaseTransactionWithCompletedAuthorizationV2
         given(transactionsViewRepository.findById(searchCriteria.transactionId))
             .willReturn(Mono.just(transactionView))
         given(
@@ -985,24 +176,28 @@ class EcommerceTransactionDataProviderTest {
     }
 
     @Test
-    fun `should map successfully transaction data into response searching by transaction id for transaction in CLOSED state`() {
+    fun `should map successfully transaction V2 data into response searching by transaction id for transaction in CLOSED state`() {
         val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
         val pageSize = 100
         val pageNumber = 0
         val transactionView =
-            TransactionTestUtils.transactionDocument(
+            TransactionTestUtilsV2.transactionDocument(
                 TransactionStatusDto.CLOSED,
                 ZonedDateTime.now()
             )
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
+        val transactionActivatedEvent = TransactionTestUtilsV2.transactionActivateEvent()
         val transactionAuthorizationRequestedEvent =
-            TransactionTestUtils.transactionAuthorizationRequestedEvent(
-                TransactionAuthorizationRequestData.PaymentGateway.XPAY
+            TransactionTestUtilsV2.transactionAuthorizationRequestedEvent(
+                TransactionAuthorizationRequestDataV2.PaymentGateway.XPAY
             )
         val transactionAuthorizationCompletedEvent =
-            TransactionTestUtils.transactionAuthorizationCompletedEvent(AuthorizationResultDto.OK)
+            TransactionTestUtilsV2.transactionAuthorizationCompletedEvent(
+                TransactionTestUtilsV2.npgTransactionGatewayAuthorizationData(
+                    OperationResultDto.EXECUTED
+                )
+            )
         val transactionClosedEvent =
-            TransactionTestUtils.transactionClosedEvent(TransactionClosureData.Outcome.OK)
+            TransactionTestUtilsV2.transactionClosedEvent(TransactionClosureDataV2.Outcome.OK)
         val events =
             listOf(
                 transactionActivatedEvent,
@@ -1010,10 +205,10 @@ class EcommerceTransactionDataProviderTest {
                 transactionAuthorizationCompletedEvent,
                 transactionClosedEvent
             )
-                as List<TransactionEvent<Any>>
+                as List<TransactionEventV2<Any>>
         val baseTransaction =
-            TransactionTestUtils.reduceEvents(*events.toTypedArray())
-                as BaseTransactionWithCompletedAuthorization
+            TransactionTestUtilsV2.reduceEvents(*events.toTypedArray())
+                as BaseTransactionWithCompletedAuthorizationV2
         given(transactionsViewRepository.findById(searchCriteria.transactionId))
             .willReturn(Mono.just(transactionView))
         given(
@@ -1099,24 +294,28 @@ class EcommerceTransactionDataProviderTest {
     }
 
     @Test
-    fun `should map successfully transaction data into response searching by transaction id for transaction in CLOSED state outcome KO`() {
+    fun `should map successfully transaction V2 data into response searching by transaction id for transaction in CLOSED state outcome KO`() {
         val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
         val pageSize = 100
         val pageNumber = 0
         val transactionView =
-            TransactionTestUtils.transactionDocument(
+            TransactionTestUtilsV2.transactionDocument(
                 TransactionStatusDto.CLOSED,
                 ZonedDateTime.now()
             )
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
+        val transactionActivatedEvent = TransactionTestUtilsV2.transactionActivateEvent()
         val transactionAuthorizationRequestedEvent =
-            TransactionTestUtils.transactionAuthorizationRequestedEvent(
-                TransactionAuthorizationRequestData.PaymentGateway.XPAY
+            TransactionTestUtilsV2.transactionAuthorizationRequestedEvent(
+                TransactionAuthorizationRequestDataV2.PaymentGateway.XPAY
             )
         val transactionAuthorizationCompletedEvent =
-            TransactionTestUtils.transactionAuthorizationCompletedEvent(AuthorizationResultDto.OK)
+            TransactionTestUtilsV2.transactionAuthorizationCompletedEvent(
+                TransactionTestUtilsV2.npgTransactionGatewayAuthorizationData(
+                    OperationResultDto.EXECUTED
+                )
+            )
         val transactionClosedEvent =
-            TransactionTestUtils.transactionClosedEvent(TransactionClosureData.Outcome.KO)
+            TransactionTestUtilsV2.transactionClosedEvent(TransactionClosureDataV2.Outcome.KO)
         val events =
             listOf(
                 transactionActivatedEvent,
@@ -1124,10 +323,10 @@ class EcommerceTransactionDataProviderTest {
                 transactionAuthorizationCompletedEvent,
                 transactionClosedEvent
             )
-                as List<TransactionEvent<Any>>
+                as List<TransactionEventV2<Any>>
         val baseTransaction =
-            TransactionTestUtils.reduceEvents(*events.toTypedArray())
-                as BaseTransactionWithCompletedAuthorization
+            TransactionTestUtilsV2.reduceEvents(*events.toTypedArray())
+                as BaseTransactionWithCompletedAuthorizationV2
         given(transactionsViewRepository.findById(searchCriteria.transactionId))
             .willReturn(Mono.just(transactionView))
         given(
@@ -1213,23 +412,27 @@ class EcommerceTransactionDataProviderTest {
     }
 
     @Test
-    fun `should map successfully transaction data into response searching by transaction id for transaction in CLOSURE_ERROR state outcome`() {
+    fun `should map successfully transaction V2 data into response searching by transaction id for transaction in CLOSURE_ERROR state outcome`() {
         val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
         val pageSize = 100
         val pageNumber = 0
         val transactionView =
-            TransactionTestUtils.transactionDocument(
+            TransactionTestUtilsV2.transactionDocument(
                 TransactionStatusDto.CLOSURE_ERROR,
                 ZonedDateTime.now()
             )
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
+        val transactionActivatedEvent = TransactionTestUtilsV2.transactionActivateEvent()
         val transactionAuthorizationRequestedEvent =
-            TransactionTestUtils.transactionAuthorizationRequestedEvent(
-                TransactionAuthorizationRequestData.PaymentGateway.XPAY
+            TransactionTestUtilsV2.transactionAuthorizationRequestedEvent(
+                TransactionAuthorizationRequestDataV2.PaymentGateway.XPAY
             )
         val transactionAuthorizationCompletedEvent =
-            TransactionTestUtils.transactionAuthorizationCompletedEvent(AuthorizationResultDto.OK)
-        val closureErrorEvent = TransactionTestUtils.transactionClosureErrorEvent()
+            TransactionTestUtilsV2.transactionAuthorizationCompletedEvent(
+                TransactionTestUtilsV2.npgTransactionGatewayAuthorizationData(
+                    OperationResultDto.EXECUTED
+                )
+            )
+        val closureErrorEvent = TransactionTestUtilsV2.transactionClosureErrorEvent()
         val events =
             listOf(
                 transactionActivatedEvent,
@@ -1237,11 +440,11 @@ class EcommerceTransactionDataProviderTest {
                 transactionAuthorizationCompletedEvent,
                 closureErrorEvent
             )
-                as List<TransactionEvent<Any>>
+                as List<TransactionEventV2<Any>>
         val baseTransaction =
-            (TransactionTestUtils.reduceEvents(*events.toTypedArray())
-                    as BaseTransactionWithClosureError)
-                .transactionAtPreviousState as BaseTransactionWithCompletedAuthorization
+            ((TransactionTestUtilsV2.reduceEvents(*events.toTypedArray())
+                    as BaseTransactionWithClosureErrorV2))
+                .transactionAtPreviousState as BaseTransactionWithCompletedAuthorizationV2
         given(transactionsViewRepository.findById(searchCriteria.transactionId))
             .willReturn(Mono.just(transactionView))
         given(
@@ -1327,22 +530,22 @@ class EcommerceTransactionDataProviderTest {
     }
 
     @Test
-    fun `should map successfully transaction data into response searching by transaction id for transaction in CANCELLATION_REQUESTED state outcome`() {
+    fun `should map successfully transaction v2 data into response searching by transaction id for transaction in CANCELLATION_REQUESTED state outcome`() {
         val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
         val pageSize = 100
         val pageNumber = 0
         val transactionView =
-            TransactionTestUtils.transactionDocument(
+            TransactionTestUtilsV2.transactionDocument(
                 TransactionStatusDto.CANCELLATION_REQUESTED,
                 ZonedDateTime.now()
             )
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
+        val transactionActivatedEvent = TransactionTestUtilsV2.transactionActivateEvent()
         val transactionCancellationRequestedEvent =
-            TransactionTestUtils.transactionUserCanceledEvent()
+            TransactionTestUtilsV2.transactionUserCanceledEvent()
         val events =
             listOf(transactionActivatedEvent, transactionCancellationRequestedEvent)
-                as List<TransactionEvent<Any>>
-        val baseTransaction = (TransactionTestUtils.reduceEvents(*events.toTypedArray()))
+                as List<TransactionEventV2<Any>>
+        val baseTransaction = (TransactionTestUtilsV2.reduceEvents(*events.toTypedArray()))
         given(transactionsViewRepository.findById(searchCriteria.transactionId))
             .willReturn(Mono.just(transactionView))
         given(
@@ -1410,21 +613,21 @@ class EcommerceTransactionDataProviderTest {
     }
 
     @Test
-    fun `should map successfully transaction data into response searching by transaction id for transaction in CANCELLATION_EXPIRED state outcome`() {
+    fun `should map successfully transaction v2 data into response searching by transaction id for transaction in CANCELLATION_EXPIRED state outcome`() {
         val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
         val pageSize = 100
         val pageNumber = 0
         val transactionView =
-            TransactionTestUtils.transactionDocument(
+            TransactionTestUtilsV2.transactionDocument(
                 TransactionStatusDto.CANCELLATION_EXPIRED,
                 ZonedDateTime.now()
             )
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
+        val transactionActivatedEvent = TransactionTestUtilsV2.transactionActivateEvent()
         val transactionCancellationRequestedEvent =
-            TransactionTestUtils.transactionUserCanceledEvent()
+            TransactionTestUtilsV2.transactionUserCanceledEvent()
         val transactionExpiredEvent =
-            TransactionTestUtils.transactionExpiredEvent(
-                TransactionTestUtils.reduceEvents(
+            TransactionTestUtilsV2.transactionExpiredEvent(
+                TransactionTestUtilsV2.reduceEvents(
                     transactionActivatedEvent,
                     transactionCancellationRequestedEvent
                 )
@@ -1435,8 +638,8 @@ class EcommerceTransactionDataProviderTest {
                 transactionCancellationRequestedEvent,
                 transactionExpiredEvent
             )
-                as List<TransactionEvent<Any>>
-        val baseTransaction = (TransactionTestUtils.reduceEvents(*events.toTypedArray()))
+                as List<TransactionEventV2<Any>>
+        val baseTransaction = (TransactionTestUtilsV2.reduceEvents(*events.toTypedArray()))
         given(transactionsViewRepository.findById(searchCriteria.transactionId))
             .willReturn(Mono.just(transactionView))
         given(
@@ -1504,20 +707,20 @@ class EcommerceTransactionDataProviderTest {
     }
 
     @Test
-    fun `should map successfully transaction data into response searching by transaction id for transaction in CANCELED state outcome`() {
+    fun `should map successfully transaction v2 data into response searching by transaction id for transaction in CANCELED state outcome`() {
         val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
         val pageSize = 100
         val pageNumber = 0
         val transactionView =
-            TransactionTestUtils.transactionDocument(
+            TransactionTestUtilsV2.transactionDocument(
                 TransactionStatusDto.CANCELED,
                 ZonedDateTime.now()
             )
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
+        val transactionActivatedEvent = TransactionTestUtilsV2.transactionActivateEvent()
         val transactionCancellationRequestedEvent =
-            TransactionTestUtils.transactionUserCanceledEvent()
+            TransactionTestUtilsV2.transactionUserCanceledEvent()
         val transactionClosedEvent =
-            TransactionTestUtils.transactionClosedEvent(TransactionClosureData.Outcome.OK)
+            TransactionTestUtilsV2.transactionClosedEvent(TransactionClosureDataV2.Outcome.OK)
 
         val events =
             listOf(
@@ -1525,8 +728,8 @@ class EcommerceTransactionDataProviderTest {
                 transactionCancellationRequestedEvent,
                 transactionClosedEvent
             )
-                as List<TransactionEvent<Any>>
-        val baseTransaction = (TransactionTestUtils.reduceEvents(*events.toTypedArray()))
+                as List<TransactionEventV2<Any>>
+        val baseTransaction = (TransactionTestUtilsV2.reduceEvents(*events.toTypedArray()))
         given(transactionsViewRepository.findById(searchCriteria.transactionId))
             .willReturn(Mono.just(transactionView))
         given(
@@ -1594,24 +797,30 @@ class EcommerceTransactionDataProviderTest {
     }
 
     @Test
-    fun `should map successfully transaction data into response searching by transaction id for transaction in UNAUTHORIZED state outcome`() {
+    fun `should map successfully transaction v2 data into response searching by transaction id for transaction in UNAUTHORIZED state outcome`() {
         val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
         val pageSize = 100
         val pageNumber = 0
         val transactionView =
-            TransactionTestUtils.transactionDocument(
+            TransactionTestUtilsV2.transactionDocument(
                 TransactionStatusDto.UNAUTHORIZED,
                 ZonedDateTime.now()
             )
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
+        val transactionActivatedEvent = TransactionTestUtilsV2.transactionActivateEvent()
         val transactionAuthorizationRequestedEvent =
-            TransactionTestUtils.transactionAuthorizationRequestedEvent(
-                TransactionAuthorizationRequestData.PaymentGateway.XPAY
+            TransactionTestUtilsV2.transactionAuthorizationRequestedEvent(
+                TransactionAuthorizationRequestDataV2.PaymentGateway.XPAY
             )
         val transactionAuthorizationCompletedEvent =
-            TransactionTestUtils.transactionAuthorizationCompletedEvent(AuthorizationResultDto.KO)
+            TransactionTestUtilsV2.transactionAuthorizationCompletedEvent(
+                TransactionTestUtilsV2.npgTransactionGatewayAuthorizationData(
+                    OperationResultDto.FAILED
+                )
+            )
         val transactionCloseFailedEvent =
-            TransactionTestUtils.transactionClosureFailedEvent(TransactionClosureData.Outcome.KO)
+            TransactionTestUtilsV2.transactionClosureFailedEvent(
+                TransactionClosureDataV2.Outcome.KO
+            )
 
         val events =
             listOf(
@@ -1620,10 +829,10 @@ class EcommerceTransactionDataProviderTest {
                 transactionAuthorizationCompletedEvent,
                 transactionCloseFailedEvent
             )
-                as List<TransactionEvent<Any>>
+                as List<TransactionEventV2<Any>>
         val baseTransaction =
-            TransactionTestUtils.reduceEvents(*events.toTypedArray())
-                as BaseTransactionWithCompletedAuthorization
+            TransactionTestUtilsV2.reduceEvents(*events.toTypedArray())
+                as BaseTransactionWithCompletedAuthorizationV2
 
         given(transactionsViewRepository.findById(searchCriteria.transactionId))
             .willReturn(Mono.just(transactionView))
@@ -1710,28 +919,34 @@ class EcommerceTransactionDataProviderTest {
     }
 
     @Test
-    fun `should map successfully transaction data into response searching by transaction id for transaction in NOTIFICATION_REQUESTED OK state outcome`() {
+    fun `should map successfully transaction v2 data into response searching by transaction id for transaction in NOTIFICATION_REQUESTED OK state outcome`() {
         val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
         val pageSize = 100
         val pageNumber = 0
         val transactionView =
-            TransactionTestUtils.transactionDocument(
+            TransactionTestUtilsV2.transactionDocument(
                 TransactionStatusDto.NOTIFICATION_REQUESTED,
                 ZonedDateTime.now()
             )
         val transactionUserReceiptData =
-            TransactionTestUtils.transactionUserReceiptData(TransactionUserReceiptData.Outcome.OK)
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
+            TransactionTestUtilsV2.transactionUserReceiptData(
+                TransactionUserReceiptDataV2.Outcome.OK
+            )
+        val transactionActivatedEvent = TransactionTestUtilsV2.transactionActivateEvent()
         val transactionAuthorizationRequestedEvent =
-            TransactionTestUtils.transactionAuthorizationRequestedEvent(
-                TransactionAuthorizationRequestData.PaymentGateway.XPAY
+            TransactionTestUtilsV2.transactionAuthorizationRequestedEvent(
+                TransactionAuthorizationRequestDataV2.PaymentGateway.XPAY
             )
         val transactionAuthorizationCompletedEvent =
-            TransactionTestUtils.transactionAuthorizationCompletedEvent(AuthorizationResultDto.OK)
+            TransactionTestUtilsV2.transactionAuthorizationCompletedEvent(
+                TransactionTestUtilsV2.npgTransactionGatewayAuthorizationData(
+                    OperationResultDto.EXECUTED
+                )
+            )
         val transactionClosedEvent =
-            TransactionTestUtils.transactionClosedEvent(TransactionClosureData.Outcome.OK)
+            TransactionTestUtilsV2.transactionClosedEvent(TransactionClosureDataV2.Outcome.OK)
         val transactionNotificationRequestedEvent =
-            TransactionTestUtils.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
+            TransactionTestUtilsV2.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
 
         val events =
             listOf(
@@ -1741,10 +956,10 @@ class EcommerceTransactionDataProviderTest {
                 transactionClosedEvent,
                 transactionNotificationRequestedEvent
             )
-                as List<TransactionEvent<Any>>
+                as List<TransactionEventV2<Any>>
         val baseTransaction =
-            TransactionTestUtils.reduceEvents(*events.toTypedArray())
-                as BaseTransactionWithRequestedUserReceipt
+            TransactionTestUtilsV2.reduceEvents(*events.toTypedArray())
+                as BaseTransactionWithRequestedUserReceiptV2
 
         given(transactionsViewRepository.findById(searchCriteria.transactionId))
             .willReturn(Mono.just(transactionView))
@@ -1839,23 +1054,29 @@ class EcommerceTransactionDataProviderTest {
         val pageSize = 100
         val pageNumber = 0
         val transactionView =
-            TransactionTestUtils.transactionDocument(
+            TransactionTestUtilsV2.transactionDocument(
                 TransactionStatusDto.NOTIFICATION_REQUESTED,
                 ZonedDateTime.now()
             )
         val transactionUserReceiptData =
-            TransactionTestUtils.transactionUserReceiptData(TransactionUserReceiptData.Outcome.KO)
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
+            TransactionTestUtilsV2.transactionUserReceiptData(
+                TransactionUserReceiptDataV2.Outcome.KO
+            )
+        val transactionActivatedEvent = TransactionTestUtilsV2.transactionActivateEvent()
         val transactionAuthorizationRequestedEvent =
-            TransactionTestUtils.transactionAuthorizationRequestedEvent(
-                TransactionAuthorizationRequestData.PaymentGateway.XPAY
+            TransactionTestUtilsV2.transactionAuthorizationRequestedEvent(
+                TransactionAuthorizationRequestDataV2.PaymentGateway.XPAY
             )
         val transactionAuthorizationCompletedEvent =
-            TransactionTestUtils.transactionAuthorizationCompletedEvent(AuthorizationResultDto.OK)
+            TransactionTestUtilsV2.transactionAuthorizationCompletedEvent(
+                TransactionTestUtilsV2.npgTransactionGatewayAuthorizationData(
+                    OperationResultDto.EXECUTED
+                )
+            )
         val transactionClosedEvent =
-            TransactionTestUtils.transactionClosedEvent(TransactionClosureData.Outcome.OK)
+            TransactionTestUtilsV2.transactionClosedEvent(TransactionClosureDataV2.Outcome.OK)
         val transactionNotificationRequestedEvent =
-            TransactionTestUtils.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
+            TransactionTestUtilsV2.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
 
         val events =
             listOf(
@@ -1865,10 +1086,10 @@ class EcommerceTransactionDataProviderTest {
                 transactionClosedEvent,
                 transactionNotificationRequestedEvent
             )
-                as List<TransactionEvent<Any>>
+                as List<TransactionEventV2<Any>>
         val baseTransaction =
-            TransactionTestUtils.reduceEvents(*events.toTypedArray())
-                as BaseTransactionWithRequestedUserReceipt
+            TransactionTestUtilsV2.reduceEvents(*events.toTypedArray())
+                as BaseTransactionWithRequestedUserReceiptV2
 
         given(transactionsViewRepository.findById(searchCriteria.transactionId))
             .willReturn(Mono.just(transactionView))
@@ -1958,30 +1179,36 @@ class EcommerceTransactionDataProviderTest {
     }
 
     @Test
-    fun `should map successfully transaction data into response searching by transaction id for transaction in NOTIFICATION_ERROR OK state outcome`() {
+    fun `should map successfully transaction v2 data into response searching by transaction id for transaction in NOTIFICATION_ERROR OK state outcome`() {
         val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
         val pageSize = 100
         val pageNumber = 0
         val transactionView =
-            TransactionTestUtils.transactionDocument(
+            TransactionTestUtilsV2.transactionDocument(
                 TransactionStatusDto.NOTIFICATION_ERROR,
                 ZonedDateTime.now()
             )
         val transactionUserReceiptData =
-            TransactionTestUtils.transactionUserReceiptData(TransactionUserReceiptData.Outcome.OK)
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
+            TransactionTestUtilsV2.transactionUserReceiptData(
+                TransactionUserReceiptDataV2.Outcome.OK
+            )
+        val transactionActivatedEvent = TransactionTestUtilsV2.transactionActivateEvent()
         val transactionAuthorizationRequestedEvent =
-            TransactionTestUtils.transactionAuthorizationRequestedEvent(
-                TransactionAuthorizationRequestData.PaymentGateway.XPAY
+            TransactionTestUtilsV2.transactionAuthorizationRequestedEvent(
+                TransactionAuthorizationRequestDataV2.PaymentGateway.XPAY
             )
         val transactionAuthorizationCompletedEvent =
-            TransactionTestUtils.transactionAuthorizationCompletedEvent(AuthorizationResultDto.OK)
+            TransactionTestUtilsV2.transactionAuthorizationCompletedEvent(
+                TransactionTestUtilsV2.npgTransactionGatewayAuthorizationData(
+                    OperationResultDto.EXECUTED
+                )
+            )
         val transactionClosedEvent =
-            TransactionTestUtils.transactionClosedEvent(TransactionClosureData.Outcome.OK)
+            TransactionTestUtilsV2.transactionClosedEvent(TransactionClosureDataV2.Outcome.OK)
         val transactionNotificationRequestedEvent =
-            TransactionTestUtils.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
+            TransactionTestUtilsV2.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
         val transactionUserReceiptError =
-            TransactionTestUtils.transactionUserReceiptAddErrorEvent(transactionUserReceiptData)
+            TransactionTestUtilsV2.transactionUserReceiptAddErrorEvent(transactionUserReceiptData)
 
         val events =
             listOf(
@@ -1992,10 +1219,10 @@ class EcommerceTransactionDataProviderTest {
                 transactionNotificationRequestedEvent,
                 transactionUserReceiptError
             )
-                as List<TransactionEvent<Any>>
+                as List<TransactionEventV2<Any>>
         val baseTransaction =
-            TransactionTestUtils.reduceEvents(*events.toTypedArray())
-                as BaseTransactionWithRequestedUserReceipt
+            TransactionTestUtilsV2.reduceEvents(*events.toTypedArray())
+                as BaseTransactionWithRequestedUserReceiptV2
 
         given(transactionsViewRepository.findById(searchCriteria.transactionId))
             .willReturn(Mono.just(transactionView))
@@ -2085,30 +1312,36 @@ class EcommerceTransactionDataProviderTest {
     }
 
     @Test
-    fun `should map successfully transaction data into response searching by transaction id for transaction in NOTIFICATION_ERROR KO state outcome`() {
+    fun `should map successfully transaction v2 data into response searching by transaction id for transaction in NOTIFICATION_ERROR KO state outcome`() {
         val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
         val pageSize = 100
         val pageNumber = 0
         val transactionView =
-            TransactionTestUtils.transactionDocument(
+            TransactionTestUtilsV2.transactionDocument(
                 TransactionStatusDto.NOTIFICATION_ERROR,
                 ZonedDateTime.now()
             )
         val transactionUserReceiptData =
-            TransactionTestUtils.transactionUserReceiptData(TransactionUserReceiptData.Outcome.KO)
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
+            TransactionTestUtilsV2.transactionUserReceiptData(
+                TransactionUserReceiptDataV2.Outcome.KO
+            )
+        val transactionActivatedEvent = TransactionTestUtilsV2.transactionActivateEvent()
         val transactionAuthorizationRequestedEvent =
-            TransactionTestUtils.transactionAuthorizationRequestedEvent(
-                TransactionAuthorizationRequestData.PaymentGateway.XPAY
+            TransactionTestUtilsV2.transactionAuthorizationRequestedEvent(
+                TransactionAuthorizationRequestDataV2.PaymentGateway.XPAY
             )
         val transactionAuthorizationCompletedEvent =
-            TransactionTestUtils.transactionAuthorizationCompletedEvent(AuthorizationResultDto.OK)
+            TransactionTestUtilsV2.transactionAuthorizationCompletedEvent(
+                TransactionTestUtilsV2.npgTransactionGatewayAuthorizationData(
+                    OperationResultDto.EXECUTED
+                )
+            )
         val transactionClosedEvent =
-            TransactionTestUtils.transactionClosedEvent(TransactionClosureData.Outcome.OK)
+            TransactionTestUtilsV2.transactionClosedEvent(TransactionClosureDataV2.Outcome.OK)
         val transactionNotificationRequestedEvent =
-            TransactionTestUtils.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
+            TransactionTestUtilsV2.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
         val transactionUserReceiptError =
-            TransactionTestUtils.transactionUserReceiptAddErrorEvent(transactionUserReceiptData)
+            TransactionTestUtilsV2.transactionUserReceiptAddErrorEvent(transactionUserReceiptData)
 
         val events =
             listOf(
@@ -2119,10 +1352,10 @@ class EcommerceTransactionDataProviderTest {
                 transactionNotificationRequestedEvent,
                 transactionUserReceiptError
             )
-                as List<TransactionEvent<Any>>
+                as List<TransactionEventV2<Any>>
         val baseTransaction =
-            TransactionTestUtils.reduceEvents(*events.toTypedArray())
-                as BaseTransactionWithRequestedUserReceipt
+            TransactionTestUtilsV2.reduceEvents(*events.toTypedArray())
+                as BaseTransactionWithRequestedUserReceiptV2
 
         given(transactionsViewRepository.findById(searchCriteria.transactionId))
             .willReturn(Mono.just(transactionView))
@@ -2212,30 +1445,36 @@ class EcommerceTransactionDataProviderTest {
     }
 
     @Test
-    fun `should map successfully transaction data into response searching by transaction id for transaction in NOTIFIED_KO state outcome`() {
+    fun `should map successfully transaction V2 data into response searching by transaction id for transaction in NOTIFIED_KO state outcome`() {
         val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
         val pageSize = 100
         val pageNumber = 0
         val transactionView =
-            TransactionTestUtils.transactionDocument(
+            TransactionTestUtilsV2.transactionDocument(
                 TransactionStatusDto.NOTIFIED_KO,
                 ZonedDateTime.now()
             )
         val transactionUserReceiptData =
-            TransactionTestUtils.transactionUserReceiptData(TransactionUserReceiptData.Outcome.KO)
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
+            TransactionTestUtilsV2.transactionUserReceiptData(
+                TransactionUserReceiptDataV2.Outcome.KO
+            )
+        val transactionActivatedEvent = TransactionTestUtilsV2.transactionActivateEvent()
         val transactionAuthorizationRequestedEvent =
-            TransactionTestUtils.transactionAuthorizationRequestedEvent(
-                TransactionAuthorizationRequestData.PaymentGateway.XPAY
+            TransactionTestUtilsV2.transactionAuthorizationRequestedEvent(
+                TransactionAuthorizationRequestDataV2.PaymentGateway.XPAY
             )
         val transactionAuthorizationCompletedEvent =
-            TransactionTestUtils.transactionAuthorizationCompletedEvent(AuthorizationResultDto.OK)
+            TransactionTestUtilsV2.transactionAuthorizationCompletedEvent(
+                TransactionTestUtilsV2.npgTransactionGatewayAuthorizationData(
+                    OperationResultDto.EXECUTED
+                )
+            )
         val transactionClosedEvent =
-            TransactionTestUtils.transactionClosedEvent(TransactionClosureData.Outcome.OK)
+            TransactionTestUtilsV2.transactionClosedEvent(TransactionClosureDataV2.Outcome.OK)
         val transactionNotificationRequestedEvent =
-            TransactionTestUtils.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
+            TransactionTestUtilsV2.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
         val transactionUserReceiptError =
-            TransactionTestUtils.transactionUserReceiptAddedEvent(transactionUserReceiptData)
+            TransactionTestUtilsV2.transactionUserReceiptAddedEvent(transactionUserReceiptData)
 
         val events =
             listOf(
@@ -2246,10 +1485,10 @@ class EcommerceTransactionDataProviderTest {
                 transactionNotificationRequestedEvent,
                 transactionUserReceiptError
             )
-                as List<TransactionEvent<Any>>
+                as List<TransactionEventV2<Any>>
         val baseTransaction =
-            TransactionTestUtils.reduceEvents(*events.toTypedArray())
-                as BaseTransactionWithRequestedUserReceipt
+            TransactionTestUtilsV2.reduceEvents(*events.toTypedArray())
+                as BaseTransactionWithRequestedUserReceiptV2
 
         given(transactionsViewRepository.findById(searchCriteria.transactionId))
             .willReturn(Mono.just(transactionView))
@@ -2344,28 +1583,34 @@ class EcommerceTransactionDataProviderTest {
         val pageSize = 100
         val pageNumber = 0
         val transactionView =
-            TransactionTestUtils.transactionDocument(
+            TransactionTestUtilsV2.transactionDocument(
                 TransactionStatusDto.REFUND_REQUESTED,
                 ZonedDateTime.now()
             )
         val transactionUserReceiptData =
-            TransactionTestUtils.transactionUserReceiptData(TransactionUserReceiptData.Outcome.KO)
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
+            TransactionTestUtilsV2.transactionUserReceiptData(
+                TransactionUserReceiptDataV2.Outcome.KO
+            )
+        val transactionActivatedEvent = TransactionTestUtilsV2.transactionActivateEvent()
         val transactionAuthorizationRequestedEvent =
-            TransactionTestUtils.transactionAuthorizationRequestedEvent(
-                TransactionAuthorizationRequestData.PaymentGateway.XPAY
+            TransactionTestUtilsV2.transactionAuthorizationRequestedEvent(
+                TransactionAuthorizationRequestDataV2.PaymentGateway.XPAY
             )
         val transactionAuthorizationCompletedEvent =
-            TransactionTestUtils.transactionAuthorizationCompletedEvent(AuthorizationResultDto.OK)
+            TransactionTestUtilsV2.transactionAuthorizationCompletedEvent(
+                TransactionTestUtilsV2.npgTransactionGatewayAuthorizationData(
+                    OperationResultDto.EXECUTED
+                )
+            )
         val transactionClosedEvent =
-            TransactionTestUtils.transactionClosedEvent(TransactionClosureData.Outcome.OK)
+            TransactionTestUtilsV2.transactionClosedEvent(TransactionClosureDataV2.Outcome.OK)
         val transactionNotificationRequestedEvent =
-            TransactionTestUtils.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
+            TransactionTestUtilsV2.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
         val transactionUserReceiptError =
-            TransactionTestUtils.transactionUserReceiptAddedEvent(transactionUserReceiptData)
+            TransactionTestUtilsV2.transactionUserReceiptAddedEvent(transactionUserReceiptData)
         val transactionRefundRequestedEvent =
-            TransactionTestUtils.transactionRefundRequestedEvent(
-                TransactionTestUtils.reduceEvents(
+            TransactionTestUtilsV2.transactionRefundRequestedEvent(
+                TransactionTestUtilsV2.reduceEvents(
                     transactionActivatedEvent,
                     transactionAuthorizationRequestedEvent,
                     transactionAuthorizationCompletedEvent,
@@ -2385,11 +1630,11 @@ class EcommerceTransactionDataProviderTest {
                 transactionUserReceiptError,
                 transactionRefundRequestedEvent
             )
-                as List<TransactionEvent<Any>>
+                as List<TransactionEventV2<Any>>
         val baseTransaction =
-            (TransactionTestUtils.reduceEvents(*events.toTypedArray())
-                    as BaseTransactionWithRefundRequested)
-                .transactionAtPreviousState as BaseTransactionWithRequestedUserReceipt
+            (TransactionTestUtilsV2.reduceEvents(*events.toTypedArray())
+                    as BaseTransactionWithRefundRequestedV2)
+                .transactionAtPreviousState as BaseTransactionWithRequestedUserReceiptV2
 
         given(transactionsViewRepository.findById(searchCriteria.transactionId))
             .willReturn(Mono.just(transactionView))
@@ -2479,33 +1724,39 @@ class EcommerceTransactionDataProviderTest {
     }
 
     @Test
-    fun `should map successfully transaction data into response searching by transaction id for transaction in REFUND_ERROR state outcome`() {
+    fun `should map successfully transaction V2 data into response searching by transaction id for transaction in REFUND_ERROR state outcome`() {
         val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
         val pageSize = 100
         val pageNumber = 0
         val transactionView =
-            TransactionTestUtils.transactionDocument(
+            TransactionTestUtilsV2.transactionDocument(
                 TransactionStatusDto.REFUND_ERROR,
                 ZonedDateTime.now()
             )
         val transactionUserReceiptData =
-            TransactionTestUtils.transactionUserReceiptData(TransactionUserReceiptData.Outcome.KO)
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
+            TransactionTestUtilsV2.transactionUserReceiptData(
+                TransactionUserReceiptDataV2.Outcome.KO
+            )
+        val transactionActivatedEvent = TransactionTestUtilsV2.transactionActivateEvent()
         val transactionAuthorizationRequestedEvent =
-            TransactionTestUtils.transactionAuthorizationRequestedEvent(
-                TransactionAuthorizationRequestData.PaymentGateway.XPAY
+            TransactionTestUtilsV2.transactionAuthorizationRequestedEvent(
+                TransactionAuthorizationRequestDataV2.PaymentGateway.XPAY
             )
         val transactionAuthorizationCompletedEvent =
-            TransactionTestUtils.transactionAuthorizationCompletedEvent(AuthorizationResultDto.OK)
+            TransactionTestUtilsV2.transactionAuthorizationCompletedEvent(
+                TransactionTestUtilsV2.npgTransactionGatewayAuthorizationData(
+                    OperationResultDto.AUTHORIZED
+                )
+            )
         val transactionClosedEvent =
-            TransactionTestUtils.transactionClosedEvent(TransactionClosureData.Outcome.OK)
+            TransactionTestUtilsV2.transactionClosedEvent(TransactionClosureDataV2.Outcome.OK)
         val transactionNotificationRequestedEvent =
-            TransactionTestUtils.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
+            TransactionTestUtilsV2.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
         val transactionUserReceiptError =
-            TransactionTestUtils.transactionUserReceiptAddedEvent(transactionUserReceiptData)
+            TransactionTestUtilsV2.transactionUserReceiptAddedEvent(transactionUserReceiptData)
         val transactionRefundRequestedEvent =
-            TransactionTestUtils.transactionRefundRequestedEvent(
-                TransactionTestUtils.reduceEvents(
+            TransactionTestUtilsV2.transactionRefundRequestedEvent(
+                TransactionTestUtilsV2.reduceEvents(
                     transactionActivatedEvent,
                     transactionAuthorizationRequestedEvent,
                     transactionAuthorizationCompletedEvent,
@@ -2515,8 +1766,8 @@ class EcommerceTransactionDataProviderTest {
                 )
             )
         val transactionRefundErrorEvent =
-            TransactionTestUtils.transactionRefundErrorEvent(
-                TransactionTestUtils.reduceEvents(
+            TransactionTestUtilsV2.transactionRefundErrorEvent(
+                TransactionTestUtilsV2.reduceEvents(
                     transactionActivatedEvent,
                     transactionAuthorizationRequestedEvent,
                     transactionAuthorizationCompletedEvent,
@@ -2537,18 +1788,18 @@ class EcommerceTransactionDataProviderTest {
                 transactionRefundRequestedEvent,
                 transactionRefundErrorEvent
             )
-                as List<TransactionEvent<Any>>
+                as List<TransactionEventV2<Any>>
         val baseTransactionRefundError =
-            TransactionTestUtils.reduceEvents(*events.toTypedArray())
-                as BaseTransactionWithRefundRequested
+            TransactionTestUtilsV2.reduceEvents(*events.toTypedArray())
+                as BaseTransactionWithRefundRequestedV2
 
         val baseTransactionRefundRequested =
             baseTransactionRefundError.transactionAtPreviousState
-                as BaseTransactionWithRefundRequested
+                as BaseTransactionWithRefundRequestedV2
 
         val baseTransaction =
             baseTransactionRefundRequested.transactionAtPreviousState
-                as BaseTransactionWithUserReceipt
+                as BaseTransactionWithUserReceiptV2
         given(transactionsViewRepository.findById(searchCriteria.transactionId))
             .willReturn(Mono.just(transactionView))
         given(
@@ -2637,33 +1888,39 @@ class EcommerceTransactionDataProviderTest {
     }
 
     @Test
-    fun `should map successfully transaction data into response searching by transaction id for transaction in REFUNDED state outcome`() {
+    fun `should map successfully transaction v2 data into response searching by transaction id for transaction in REFUNDED state outcome`() {
         val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
         val pageSize = 100
         val pageNumber = 0
         val transactionView =
-            TransactionTestUtils.transactionDocument(
+            TransactionTestUtilsV2.transactionDocument(
                 TransactionStatusDto.REFUNDED,
                 ZonedDateTime.now()
             )
         val transactionUserReceiptData =
-            TransactionTestUtils.transactionUserReceiptData(TransactionUserReceiptData.Outcome.KO)
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
+            TransactionTestUtilsV2.transactionUserReceiptData(
+                TransactionUserReceiptDataV2.Outcome.KO
+            )
+        val transactionActivatedEvent = TransactionTestUtilsV2.transactionActivateEvent()
         val transactionAuthorizationRequestedEvent =
-            TransactionTestUtils.transactionAuthorizationRequestedEvent(
-                TransactionAuthorizationRequestData.PaymentGateway.XPAY
+            TransactionTestUtilsV2.transactionAuthorizationRequestedEvent(
+                TransactionAuthorizationRequestDataV2.PaymentGateway.XPAY
             )
         val transactionAuthorizationCompletedEvent =
-            TransactionTestUtils.transactionAuthorizationCompletedEvent(AuthorizationResultDto.OK)
+            TransactionTestUtilsV2.transactionAuthorizationCompletedEvent(
+                TransactionTestUtilsV2.npgTransactionGatewayAuthorizationData(
+                    OperationResultDto.EXECUTED
+                )
+            )
         val transactionClosedEvent =
-            TransactionTestUtils.transactionClosedEvent(TransactionClosureData.Outcome.OK)
+            TransactionTestUtilsV2.transactionClosedEvent(TransactionClosureDataV2.Outcome.OK)
         val transactionNotificationRequestedEvent =
-            TransactionTestUtils.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
+            TransactionTestUtilsV2.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
         val transactionUserReceiptError =
-            TransactionTestUtils.transactionUserReceiptAddedEvent(transactionUserReceiptData)
+            TransactionTestUtilsV2.transactionUserReceiptAddedEvent(transactionUserReceiptData)
         val transactionRefundRequestedEvent =
-            TransactionTestUtils.transactionRefundRequestedEvent(
-                TransactionTestUtils.reduceEvents(
+            TransactionTestUtilsV2.transactionRefundRequestedEvent(
+                TransactionTestUtilsV2.reduceEvents(
                     transactionActivatedEvent,
                     transactionAuthorizationRequestedEvent,
                     transactionAuthorizationCompletedEvent,
@@ -2673,8 +1930,8 @@ class EcommerceTransactionDataProviderTest {
                 )
             )
         val transactionRefundErrorEvent =
-            TransactionTestUtils.transactionRefundErrorEvent(
-                TransactionTestUtils.reduceEvents(
+            TransactionTestUtilsV2.transactionRefundErrorEvent(
+                TransactionTestUtilsV2.reduceEvents(
                     transactionActivatedEvent,
                     transactionAuthorizationRequestedEvent,
                     transactionAuthorizationCompletedEvent,
@@ -2685,8 +1942,8 @@ class EcommerceTransactionDataProviderTest {
                 )
             )
         val transactionRefundedEvent =
-            TransactionTestUtils.transactionRefundedEvent(
-                TransactionTestUtils.reduceEvents(
+            TransactionTestUtilsV2.transactionRefundedEvent(
+                TransactionTestUtilsV2.reduceEvents(
                     transactionActivatedEvent,
                     transactionAuthorizationRequestedEvent,
                     transactionAuthorizationCompletedEvent,
@@ -2709,21 +1966,22 @@ class EcommerceTransactionDataProviderTest {
                 transactionRefundErrorEvent,
                 transactionRefundedEvent
             )
-                as List<TransactionEvent<Any>>
+                as List<TransactionEventV2<Any>>
         val baseTransactionRefunded =
-            TransactionTestUtils.reduceEvents(*events.toTypedArray())
-                as BaseTransactionWithRefundRequested
+            TransactionTestUtilsV2.reduceEvents(*events.toTypedArray())
+                as BaseTransactionWithRefundRequestedV2
 
         val baseTransactionRefundError =
-            baseTransactionRefunded.transactionAtPreviousState as BaseTransactionWithRefundRequested
+            baseTransactionRefunded.transactionAtPreviousState
+                as BaseTransactionWithRefundRequestedV2
 
         val baseTransactionRefundRequested =
             baseTransactionRefundError.transactionAtPreviousState
-                as BaseTransactionWithRefundRequested
+                as BaseTransactionWithRefundRequestedV2
 
         val baseTransaction =
             baseTransactionRefundRequested.transactionAtPreviousState
-                as BaseTransactionWithUserReceipt
+                as BaseTransactionWithUserReceiptV2
         given(transactionsViewRepository.findById(searchCriteria.transactionId))
             .willReturn(Mono.just(transactionView))
         given(
@@ -2812,33 +2070,39 @@ class EcommerceTransactionDataProviderTest {
     }
 
     @Test
-    fun `should map successfully transaction data into response searching by transaction id for transaction in REFUNDED FROM EXPIRED state outcome`() {
+    fun `should map successfully transaction v2 data into response searching by transaction id for transaction in REFUNDED FROM EXPIRED state outcome`() {
         val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
         val pageSize = 100
         val pageNumber = 0
         val transactionView =
-            TransactionTestUtils.transactionDocument(
+            TransactionTestUtilsV2.transactionDocument(
                 TransactionStatusDto.REFUNDED,
                 ZonedDateTime.now()
             )
         val transactionUserReceiptData =
-            TransactionTestUtils.transactionUserReceiptData(TransactionUserReceiptData.Outcome.KO)
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
+            TransactionTestUtilsV2.transactionUserReceiptData(
+                TransactionUserReceiptDataV2.Outcome.KO
+            )
+        val transactionActivatedEvent = TransactionTestUtilsV2.transactionActivateEvent()
         val transactionAuthorizationRequestedEvent =
-            TransactionTestUtils.transactionAuthorizationRequestedEvent(
-                TransactionAuthorizationRequestData.PaymentGateway.XPAY
+            TransactionTestUtilsV2.transactionAuthorizationRequestedEvent(
+                TransactionAuthorizationRequestDataV2.PaymentGateway.XPAY
             )
         val transactionAuthorizationCompletedEvent =
-            TransactionTestUtils.transactionAuthorizationCompletedEvent(AuthorizationResultDto.OK)
+            TransactionTestUtilsV2.transactionAuthorizationCompletedEvent(
+                TransactionTestUtilsV2.npgTransactionGatewayAuthorizationData(
+                    OperationResultDto.EXECUTED
+                )
+            )
         val transactionClosedEvent =
-            TransactionTestUtils.transactionClosedEvent(TransactionClosureData.Outcome.OK)
+            TransactionTestUtilsV2.transactionClosedEvent(TransactionClosureDataV2.Outcome.OK)
         val transactionNotificationRequestedEvent =
-            TransactionTestUtils.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
+            TransactionTestUtilsV2.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
         val transactionUserReceiptError =
-            TransactionTestUtils.transactionUserReceiptAddedEvent(transactionUserReceiptData)
+            TransactionTestUtilsV2.transactionUserReceiptAddedEvent(transactionUserReceiptData)
         val transactionExpiredEvent =
-            TransactionTestUtils.transactionExpiredEvent(
-                TransactionTestUtils.reduceEvents(
+            TransactionTestUtilsV2.transactionExpiredEvent(
+                TransactionTestUtilsV2.reduceEvents(
                     transactionActivatedEvent,
                     transactionAuthorizationRequestedEvent,
                     transactionAuthorizationCompletedEvent,
@@ -2848,8 +2112,8 @@ class EcommerceTransactionDataProviderTest {
                 )
             )
         val transactionRefundRequestedEvent =
-            TransactionTestUtils.transactionRefundRequestedEvent(
-                TransactionTestUtils.reduceEvents(
+            TransactionTestUtilsV2.transactionRefundRequestedEvent(
+                TransactionTestUtilsV2.reduceEvents(
                     transactionActivatedEvent,
                     transactionAuthorizationRequestedEvent,
                     transactionAuthorizationCompletedEvent,
@@ -2860,8 +2124,8 @@ class EcommerceTransactionDataProviderTest {
                 )
             )
         val transactionRefundErrorEvent =
-            TransactionTestUtils.transactionRefundErrorEvent(
-                TransactionTestUtils.reduceEvents(
+            TransactionTestUtilsV2.transactionRefundErrorEvent(
+                TransactionTestUtilsV2.reduceEvents(
                     transactionActivatedEvent,
                     transactionAuthorizationRequestedEvent,
                     transactionAuthorizationCompletedEvent,
@@ -2873,8 +2137,8 @@ class EcommerceTransactionDataProviderTest {
                 )
             )
         val transactionRefundedEvent =
-            TransactionTestUtils.transactionRefundedEvent(
-                TransactionTestUtils.reduceEvents(
+            TransactionTestUtilsV2.transactionRefundedEvent(
+                TransactionTestUtilsV2.reduceEvents(
                     transactionActivatedEvent,
                     transactionAuthorizationRequestedEvent,
                     transactionAuthorizationCompletedEvent,
@@ -2899,21 +2163,22 @@ class EcommerceTransactionDataProviderTest {
                 transactionRefundErrorEvent,
                 transactionRefundedEvent
             )
-                as List<TransactionEvent<Any>>
+                as List<TransactionEventV2<Any>>
         val baseTransactionRefunded =
-            TransactionTestUtils.reduceEvents(*events.toTypedArray())
-                as BaseTransactionWithRefundRequested
+            TransactionTestUtilsV2.reduceEvents(*events.toTypedArray())
+                as BaseTransactionWithRefundRequestedV2
 
         val baseTransactionRefundError =
-            baseTransactionRefunded.transactionAtPreviousState as BaseTransactionWithRefundRequested
+            baseTransactionRefunded.transactionAtPreviousState
+                as BaseTransactionWithRefundRequestedV2
 
         val baseTransactionRefundRequested =
             baseTransactionRefundError.transactionAtPreviousState
-                as BaseTransactionWithRefundRequested
+                as BaseTransactionWithRefundRequestedV2
 
         val baseTransaction =
             baseTransactionRefundRequested.transactionAtPreviousState
-                as BaseTransactionWithUserReceipt
+                as BaseTransactionWithUserReceiptV2
         given(transactionsViewRepository.findById(searchCriteria.transactionId))
             .willReturn(Mono.just(transactionView))
         given(
@@ -3002,33 +2267,39 @@ class EcommerceTransactionDataProviderTest {
     }
 
     @Test
-    fun `should map successfully transaction data into response searching by transaction id for transaction in EXPIRED from NOTIFICATION_ERROR OK state outcome`() {
+    fun `should map successfully transaction v2 data into response searching by transaction id for transaction in EXPIRED from NOTIFICATION_ERROR OK state outcome`() {
         val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
         val pageSize = 100
         val pageNumber = 0
         val transactionView =
-            TransactionTestUtils.transactionDocument(
+            TransactionTestUtilsV2.transactionDocument(
                 TransactionStatusDto.EXPIRED,
                 ZonedDateTime.now()
             )
         val transactionUserReceiptData =
-            TransactionTestUtils.transactionUserReceiptData(TransactionUserReceiptData.Outcome.OK)
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
+            TransactionTestUtilsV2.transactionUserReceiptData(
+                TransactionUserReceiptDataV2.Outcome.OK
+            )
+        val transactionActivatedEvent = TransactionTestUtilsV2.transactionActivateEvent()
         val transactionAuthorizationRequestedEvent =
-            TransactionTestUtils.transactionAuthorizationRequestedEvent(
-                TransactionAuthorizationRequestData.PaymentGateway.XPAY
+            TransactionTestUtilsV2.transactionAuthorizationRequestedEvent(
+                TransactionAuthorizationRequestDataV2.PaymentGateway.XPAY
             )
         val transactionAuthorizationCompletedEvent =
-            TransactionTestUtils.transactionAuthorizationCompletedEvent(AuthorizationResultDto.OK)
+            TransactionTestUtilsV2.transactionAuthorizationCompletedEvent(
+                TransactionTestUtilsV2.npgTransactionGatewayAuthorizationData(
+                    OperationResultDto.EXECUTED
+                )
+            )
         val transactionClosedEvent =
-            TransactionTestUtils.transactionClosedEvent(TransactionClosureData.Outcome.OK)
+            TransactionTestUtilsV2.transactionClosedEvent(TransactionClosureDataV2.Outcome.OK)
         val transactionNotificationRequestedEvent =
-            TransactionTestUtils.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
+            TransactionTestUtilsV2.transactionUserReceiptRequestedEvent(transactionUserReceiptData)
         val transactionUserReceiptError =
-            TransactionTestUtils.transactionUserReceiptAddErrorEvent(transactionUserReceiptData)
+            TransactionTestUtilsV2.transactionUserReceiptAddErrorEvent(transactionUserReceiptData)
         val transactionExpired =
-            TransactionTestUtils.transactionExpiredEvent(
-                TransactionTestUtils.reduceEvents(
+            TransactionTestUtilsV2.transactionExpiredEvent(
+                TransactionTestUtilsV2.reduceEvents(
                     transactionActivatedEvent,
                     transactionAuthorizationRequestedEvent,
                     transactionAuthorizationCompletedEvent,
@@ -3047,12 +2318,12 @@ class EcommerceTransactionDataProviderTest {
                 transactionUserReceiptError,
                 transactionExpired
             )
-                as List<TransactionEvent<Any>>
+                as List<TransactionEventV2<Any>>
         val baseTransactionExpired =
-            TransactionTestUtils.reduceEvents(*events.toTypedArray()) as BaseTransactionExpired
+            TransactionTestUtilsV2.reduceEvents(*events.toTypedArray()) as BaseTransactionExpiredV2
         val baseTransaction =
             baseTransactionExpired.transactionAtPreviousState
-                as BaseTransactionWithRequestedUserReceipt
+                as BaseTransactionWithRequestedUserReceiptV2
 
         given(transactionsViewRepository.findById(searchCriteria.transactionId))
             .willReturn(Mono.just(transactionView))
@@ -3142,24 +2413,24 @@ class EcommerceTransactionDataProviderTest {
     }
 
     @Test
-    fun `should map successfully transaction data into response searching by transaction id for transaction in EXPIRED_NOT_AUTHORIZED state`() {
+    fun `should map successfully transaction V2 data into response searching by transaction id for transaction in EXPIRED_NOT_AUTHORIZED state`() {
         val searchCriteria = HelpdeskTestUtils.buildSearchRequestByTransactionId()
         val pageSize = 100
         val pageNumber = 0
         val transactionView =
-            TransactionTestUtils.transactionDocument(
+            TransactionTestUtilsV2.transactionDocument(
                 TransactionStatusDto.EXPIRED_NOT_AUTHORIZED,
                 ZonedDateTime.now()
             )
-        val transactionActivatedEvent = TransactionTestUtils.transactionActivateEvent()
+        val transactionActivatedEvent = TransactionTestUtilsV2.transactionActivateEvent()
         val transactionExpiredEvent =
-            TransactionTestUtils.transactionExpiredEvent(
-                TransactionTestUtils.reduceEvents(transactionActivatedEvent)
+            TransactionTestUtilsV2.transactionExpiredEvent(
+                TransactionTestUtilsV2.reduceEvents(transactionActivatedEvent)
             )
         val events =
             listOf(transactionActivatedEvent, transactionExpiredEvent)
-                as List<TransactionEvent<Any>>
-        val baseTransaction = TransactionTestUtils.reduceEvents(*events.toTypedArray())
+                as List<TransactionEventV2<Any>>
+        val baseTransaction = TransactionTestUtilsV2.reduceEvents(*events.toTypedArray())
         given(transactionsViewRepository.findById(searchCriteria.transactionId))
             .willReturn(Mono.just(transactionView))
         given(
